+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 2.1.0 - 2023-10-15
+
+### Added
+
+- `Innmind\IO\Readable\Chunks::lazy()`
+- `Innmind\IO\Readable\Chunks\Lazy`
+- `Innmind\IO\Readable\Stream::lines()`
+- `Innmind\IO\Readable\Lines`
+
 ## 2.0.0 - 2023-09-17
 
 ### Added